Add shadows to cards and header
Shadows are quite bold, may want to knock those down.
I needed to add a little margin to the bottom of the
header for this or the shadow doesn't show (I presume
it ends up overridden by the Main section).
It's probably possible to fix that by changing the
Z-order, but margin seems a bit more flexible.